Adding new BalanceManager class
