+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/controllers/main/HomeController.kt
@@ -5,6 +5,7 @@ import fr.acinq.lightning.channel.*
 import fr.acinq.lightning.utils.sum
 import fr.acinq.phoenix.PhoenixBusiness
 import fr.acinq.phoenix.controllers.AppController
+import fr.acinq.phoenix.managers.BalanceManager
 import fr.acinq.phoenix.managers.PaymentsManager
 import fr.acinq.phoenix.managers.PeerManager
 import fr.acinq.phoenix.utils.extensions.calculateBalance
@@ -17,7 +18,8 @@ import org.kodein.log.LoggerFactory
 class AppHomeController(
     loggerFactory: LoggerFactory,
     private val peerManager: PeerManager,
-    private val paymentsManager: PaymentsManager
+    private val paymentsManager: PaymentsManager,
+    private val balanceManager: BalanceManager
 ) : AppController<Home.Model, Home.Intent>(
     loggerFactory = loggerFactory,
     firstModel = Home.emptyModel
@@ -25,7 +27,8 @@ class AppHomeController(
     constructor(business: PhoenixBusiness): this(
         loggerFactory = business.loggerFactory,
         peerManager = business.peerManager,
-        paymentsManager = business.paymentsManager
+        paymentsManager = business.paymentsManager,
+        balanceManager = business.balanceManager
     )
 
     private var isBoot = true
@@ -50,7 +53,7 @@ class AppHomeController(
         }
 
         launch {
-            paymentsManager.incomingSwaps.collect {
+            balanceManager.incomingSwaps.collect {
                 model {
                     copy(incomingBalance = it.values.sum().takeIf { it.msat > 0 })
                 }
