+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/managers/PaymentsManager.kt
@@ -25,15 +25,13 @@ import org.kodein.log.newLogger
 class PaymentsManager(
     private val loggerFactory: LoggerFactory,
     private val peerManager: PeerManager,
-    private val databaseManager: DatabaseManager,
-    private val nodeParamsManager: NodeParamsManager,
+    private val databaseManager: DatabaseManager
 ) : CoroutineScope by MainScope() {
 
     constructor(business: PhoenixBusiness): this(
         loggerFactory = business.loggerFactory,
         peerManager = business.peerManager,
-        databaseManager = business.databaseManager,
-        nodeParamsManager = business.nodeParamsManager,
+        databaseManager = business.databaseManager
     )
 
     private val log = newLogger(loggerFactory)
@@ -46,14 +44,6 @@ class PaymentsManager(
     val paymentsCount: StateFlow<Long> = _paymentsCount
 
     /**
-     * Flow of map of (bitcoinAddress -> amount) swap-ins.
-     * Deprecated: Replace with PeerManager.swapInWalletBalance
-     */
-    private val _incomingSwaps = MutableStateFlow<Map<String, MilliSatoshi>>(HashMap())
-    val incomingSwaps: StateFlow<Map<String, MilliSatoshi>> = _incomingSwaps
-    private var _incomingSwapsMap by _incomingSwaps
-
-    /**
      * Broadcasts the most recently completed payment since the app was launched.
      * This includes incoming & outgoing payments (both successful & failed).
      *
@@ -61,6 +51,9 @@ class PaymentsManager(
      */
     private val _lastCompletedPayment = MutableStateFlow<WalletPayment?>(null)
     val lastCompletedPayment: StateFlow<WalletPayment?> = _lastCompletedPayment
+
+    private val _inFlightOutgoingPayments = MutableStateFlow<Set<UUID>>(setOf())
+    val inFlightOutgoingPayments: StateFlow<Set<UUID>> = _inFlightOutgoingPayments
 
     /**
      * Provides a default PaymentsFetcher for use by the app.
@@ -77,9 +70,6 @@ class PaymentsManager(
     fun makePageFetcher(): PaymentsPageFetcher {
         return PaymentsPageFetcher(loggerFactory, databaseManager)
     }
-    
-    private val _inFlightOutgoingPayments = MutableStateFlow<Set<UUID>>(setOf())
-    val inFlightOutgoingPayments: StateFlow<Set<UUID>> = _inFlightOutgoingPayments
 
     init {
         launch {
@@ -136,36 +126,6 @@ class PaymentsManager(
                         removeFromInFlightOutgoingPayments(event.request.paymentId)
                     }
                     else -> Unit
-                }
-            }
-        }
-        launch {
-            nodeParamsManager.nodeParams.filterNotNull().first().nodeEvents.collect {
-                when (it) {
-                    is SwapInEvents.Requested -> {
-                        // for now, we use a placeholder address because it's not yet exposed by the lightning-kmp API
-                        _incomingSwapsMap = _incomingSwapsMap + ("foobar" to it.req.localFundingAmount.toMilliSatoshi())
-                    }
-                    is SwapInEvents.Accepted -> {
-                        log.info { "swap-in request=${it.requestId} has been accepted for funding_fee=${it.fundingFee} service_fee=${it.serviceFee}" }
-                    }
-                    is SwapInEvents.Rejected -> {
-                        log.error { "rejected swap-in for required_fee=${it.requiredFees} with error=${it.failure}" }
-                        _incomingSwapsMap = _incomingSwapsMap - "foobar"
-                    }
-                    is ChannelEvents.Creating -> {
-                        log.info { "channel=${it.state.channelId} is being created" }
-                    }
-                    is ChannelEvents.Created -> {
-                        log.info { "channel=${it.state.channelId} has been successfully created!" }
-                        _incomingSwapsMap = _incomingSwapsMap - "foobar"
-                    }
-                    is ChannelEvents.Confirmed -> {
-                        paymentsDb().updateNewChannelConfirmed(
-                            channelId = it.state.channelId,
-                            receivedAt = currentTimestampMillis()
-                        )
-                    }
                 }
             }
         }
