+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/managers/PeerManager.kt
@@ -1,23 +1,15 @@
 package fr.acinq.phoenix.managers
 
 import fr.acinq.bitcoin.ByteVector32
-import fr.acinq.lightning.MilliSatoshi
 import fr.acinq.bitcoin.Crypto
-import fr.acinq.bitcoin.Satoshi
-import fr.acinq.lightning.ChannelEvents
 import fr.acinq.lightning.blockchain.electrum.ElectrumWatcher
-import fr.acinq.lightning.blockchain.electrum.WalletState
-import fr.acinq.lightning.blockchain.electrum.WalletState.Utxo
 import fr.acinq.lightning.channel.ChannelStateWithCommitments
-import fr.acinq.lightning.channel.*
 import fr.acinq.lightning.channel.Offline
 import fr.acinq.lightning.io.Peer
 import fr.acinq.lightning.io.TcpSocket
-import fr.acinq.lightning.utils.sat
 import fr.acinq.lightning.wire.InitTlv
 import fr.acinq.lightning.wire.TlvStream
 import fr.acinq.phoenix.PhoenixBusiness
-import fr.acinq.phoenix.utils.extensions.calculateBalance
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.flow.*
@@ -48,16 +40,6 @@ class PeerManager(
     private val _peer = MutableStateFlow<Peer?>(null)
     val peerState: StateFlow<Peer?> = _peer
 
-    private val _balance = MutableStateFlow<MilliSatoshi?>(null)
-    val balance: StateFlow<MilliSatoshi?> = _balance
-
-    private val _swapInWallet = MutableStateFlow<WalletState?>(null)
-    private val _pendingReservedUtxos = MutableStateFlow<Map<ByteVector32, List<Utxo>>>(emptyMap())
-    private val _reservedUtxos = MutableStateFlow<Set<Utxo>>(emptySet())
-
-    private val _swapInWalletBalance = MutableStateFlow(WalletBalance.empty())
-    val swapInWalletBalance: StateFlow<WalletBalance> = _swapInWalletBalance
-
     init {
         launch {
             val nodeParams = nodeParamsManager.nodeParams.filterNotNull().first()
@@ -86,76 +68,6 @@ class PeerManager(
                 scope = MainScope()
             )
             _peer.value = peer
-            monitorPeer(peer)
-        }
-    }
-
-    private fun monitorPeer(peer: Peer) {
-        launch {
-            peer.channelsFlow.collect { channels ->
-                _balance.value = calculateBalance(channels)
-            }
-        }
-
-        /**
-         * What is the swap-in wallet's balance ?
-         *
-         * At first glance, it looks like we should be able to use:
-         * `peer.swapInWallet.walletStateFlow`
-         *
-         * However, this is a reflection of the balance on the Electrum server.
-         * Which means, there's a *DELAY* between:
-         * - when we consider the channel opened (+ update balance, + create IncomingPayment)
-         * - when that updated balance is bounced back from the Electrum server
-         *
-         * Even on a fast connection, that takes several seconds.
-         * And during that time, the UI is in a bad state:
-         *
-         * - the (lightning) balance has been updated to reflect the new channel
-         * - the incoming payment is reflected in the payments list
-         * - but the wallet incorrectly says "+ X sat incoming"
-         *
-         * So we work around this shortcoming by taking advantage of:
-         * - ChannelEvents.Creating
-         * - ChannelEvents.Created
-         *
-         * Which gives us the `List<Utxo>` that we can manually ignore,
-         * while we wait for the Electrum wallet to catch up.
-         */
-
-        launch {
-            peer.swapInWallet.walletStateFlow.collect { wallet ->
-                _swapInWallet.value = wallet
-                _reservedUtxos.update { it.intersect(wallet.utxos) }
-            }
-        }
-        launch {
-            peer.nodeParams.nodeEvents.collect { event ->
-                when (event) {
-                    is ChannelEvents.Creating -> {
-                        val channelId = event.state.channelId
-                        val channelUtxos = event.state.wallet.confirmedUtxos
-                        _pendingReservedUtxos.update { it.plus(channelId to channelUtxos) }
-                    }
-                    is ChannelEvents.Created -> {
-                        val channelId = event.state.channelId
-                        _pendingReservedUtxos.value[channelId]?.let { channelUtxos ->
-                            _reservedUtxos.update { it.union(channelUtxos) }
-                            _pendingReservedUtxos.update { it.minus(channelId) }
-                        }
-                    }
-                }
-            }
-        }
-        launch {
-            combine(_swapInWallet.filterNotNull(), _reservedUtxos) { swapInWallet, reservedUtxos ->
-                swapInWallet.minus(reservedUtxos)
-            }.collect { availableWallet ->
-                _swapInWalletBalance.value = WalletBalance(
-                    confirmed = availableWallet.confirmedBalance,
-                    unconfirmed = availableWallet.unconfirmedBalance
-                )
-            }
         }
     }
 
@@ -176,16 +88,5 @@ class PeerManager(
             is ChannelStateWithCommitments -> channel
             else -> null
         }
-    }
-}
-
-data class WalletBalance(
-    val confirmed: Satoshi,
-    val unconfirmed: Satoshi
-) {
-    val total get() = confirmed + unconfirmed
-
-    companion object {
-        fun empty() = WalletBalance(0.sat, 0.sat)
     }
 }
